schism: fix MIDI "all notes off" on stop
The existing code sent a sequence of:
1. all sound off
2. reset all controllers
3. all notes off

Which are all channel control messsages, and must be sent for every channel.
A channel of 0 was hard-coded for all of these, which prevented F8 from
effectively stopping MIDI notes when any channels other than the first were
in use.

This commit just sends the same sequence, for all the channels.