clippy: reply with SelectionRequest's target
In the event that we don't have the appropriate target type, the
conversion fails.  We should not be replying with the type we do have
as the target and the property None.  Rather we should reply with the
target they requested and property None.

From XSelectionEvent(3):

> The requestor member is set to the window associated with the
> requestor of the selection.  The selection member is set to the atom
> that indicates the selection.  For example, PRIMARY is used for the
> primary selection.
>
> The target member is set to the atom that indicates the converted
> type.  For example, PIXMAP is used for a pixmap.  The property
> member is set to the atom that indicates which property the result
> was stored on.
>
> If the conversion failed, the property member is set to None.  The
> time member is set to the time the conversion took place and can be
> a timestamp or CurrentTime.

It's slightly ambiguous, but this is Xlib after all.

With the code as-is, when I attempt to paste into xterm from schism,
xterm asks for a conversion to compound text.  When schism replies
with a string type and property none, xterm seems to completely ignore
this notification as unrelated, and enters into a strange delayed
process of slowly retrying selection requests of different atoms.

With only the previous commit in place, this slow retry process
eventually succeeds after xterm finally gives up on the conversion.
With this commit as well, xterm is immediately successful in
retrieving the cut buffer on Shift+Insert.