New instrument command: delete-with-preserve (#196)
This commit adds a new command to the instrument page:

    Delete instrument & all related unused samples

This command is bound to Alt-Shift-D.

When executed, this command deletes the currently selected instrument
and all related samples that are not used by any other instrument.  This
is similar to Alt-D, except it does not delete samples used by other
instruments.

Example: If instruments X and Y both use sample S, you can delete one
instrument without breaking the other.