fmt/s3m: clamp mixing volume on save
In S3M this field has only seven bits, as the top bit is used by
the stereo flag.

Without the clamp, a max volume save of 0x80 only sets the stereo
bit with a MasterVolume of zero.

I didn't bother trying to fill out the 0-128 range on import, as-is
the .s3m import will only reach mixing volume up to 127.

Discovered and reported by Cranksy @ http://cranksy.net, thanks!