*: DO NOT MERGE: JUNK DRAWER "MAKE IT BUILD" COMMIT
This commit is just whatever was uncommitted in my SDL2 port WIP tree,
assume it's of garbage quality and full of hacks just to get things
compiling again.

Surprisingly it seems to actually produce a runnable SDL2 port here.

Without this commit, the tree doesn't build.  But I'm sure this
commit encompasses a bunch of unrelated junk, not strictly what's
necessary to make things build again.