video: Fix OOB memory access on 24-bit displays
SchismTracker will write 32-bits at a time to a 24-bit display when
scaling, which means it can overflow the end of the surface its writing
into. This usually isn't a problem: many implementations will be
tolerant of writing an extra byte or two past the end of the surface,
but sdl12-compat, for instance, isn't. See:
https://github.com/libsdl-org/sdl12-compat/issues/183