rocket: minimally wire up rocket to build
This introduces a new '--with-rocket' configure flag, the default
when unspecified is for rocket support to not be included.

No integration code exists yet, but this makes available the
necessary parts of the GNU Rocket library to start on that.

There probably needs to be more work on configure/build side
since the Rocket code has some platform-specific ifdefs of its
own we're not currently managing.  I'd rather not run their
full-blown build if possible, considering how small the library
bits are to integrate.  We should really be able to get away with
more or less what's done in this commit.