Set SDL_HINT_GRAB_KEYBOARD to "1", fix #373
`SDL_SetWindowGrab()` only grabs mouse in SDL2, unless this hint is set to 1. Fixes keyboard grabbing with Ctrl-D

If we wanted to ONLY grab keyboard, and not grab mouse, we could use
`SDL_SetWindowKeyboardGrab()` instead.

Note that under wayland specifically, this will only work with the
`wayland` video driver. You need to force that manually with the
`SDL_VIDEODRIVER=wayland` env var, but thats just because its not
considered stable in SDL2. I think SDL3 is trying to enable it by
default. Not really relevant to schism I think.