Fix out of bounds memory access rendering samples
Right now the sample renderer will perform out of bounds reads after
the end of the sample data. As far as I can tell, this always happens.
Usually this doesn't result in a crash and will at worst result in a
rendering error, but with particularly large samples it will read so
far past the end of the sample data that it can generate a segfault.

This fixes the rendering code to stay in bounds, with special handling
of 1-length samples. I've tested this on my long samples that
generated a crash 100% of the time, as well as on 1-length 8-bit mono
and 16-bit stereo/mono samples.