sample-view.c: cleanup sample drawing code
related to #376

alright, a bit of explaining: before this commit, the sample drawing code

unintentionally would skip past the desired length, because it would increase the step

before actually checking if it was legitimate or not. this results in weird behavior,

especially in large samples.

I'm not entirely sure if this fixes it or not, but the code sure as hell looks a lot

cleaner now.