player/fx: fix stupid regression in d1ba60e
I never actually checked any test cases after making this change,
and assumed my code was right (it most definitely is not)

I spent a couple days looking through my code and found out that
the New Note Action is actually triggered by whether the sample is
playing *at all*, not whether it's muted or not. (this probably
should have been obvious, but I wasn't looking in the right places)

what a mess.