midi/jack: much more progress
we now have actual working ports!

(well, sort of. We can't tell which port is sending which event
because of how jack-midi works but that's not really necessary
anyway and is only there for pretty GUI stuff)

for some reason I can't for the life of me get the midi write code
working. Something is either really wrong with what I'm writing OR
jack is just eating our midi data. either way, doesn't work :)

related: #144