fmt: permit saving instruments to file formats other than IT
also I removed all of the remaining IT-specific junk out of
audio_loadsave.c and moved it to its correct home in fmt/it.c.

also I combined the ITI loaders so now they're using the same
code in both the it instrument loader and it song loader.

also there's a new keybind (Alt-T, same as the sample list) that
allows you to select from a list of formats to export in. It only
has IT for now tho.