Do not fill DMOZ directory for save dialog when initializing Schism Tracker
It should not be necessary to initialize the directory contents when initializing the save/export pages on startup.
The other directory browser pages don't do this either. This call resulted the working directory to be scanned for samples and modules, which could slow down Schism Tracker startup considerably in particular if Media Foundation support is active.