fmt: edit the loader functions to use slurp
this is just one step in removing the reliance on memory mapping...
there are still lots of lines that require memory mapping and hence
I've changed those to either use the slurp members or use the slurp
functions.

Note that I'm not entirely REMOVING memory mapping, because that's
kind of dumb. However it is useful to add a real not-so-stupid
way of accessing files when mmap and MapViewOfFile aren't available;
i.e. when not on posix or win32 (wii?? maybe? dunno really.)