*: megacommit, sorry...
This commit mostly contains fixes for mac os and fixes a super edge case
in UTF-8 character drawing where it would go out of bounds when decoding
failed

This also removes the dependency on the POSIX function mkstemp(), replaced
by our own function which uses only stuff provided by the C standard
library.
Note that it may be possible to use mkstemp(), followed by a fchmod() to
fix the permissions on the file.