fmt: store 32-bit version info in S3M and IT files
We don't have to edit the IT code at all, since it already saved
the reserved field as a 32-bit integer.

S3M splits the number between the high 2 and low 2 bytes of the
reserved field. The high 2 come after the edit timer, and the low 2 come
after the edit timer. This is so we can retain compatibility
with S3M files made with git versions of schism, as well as
keeping players and other trackers that know about the edit
timer from having to account for the value moving.

This results in the version info overflow being moved from 5 April 2189
to 19 November 11761230, both of which are absolutely ridiculous.