version: remove usage of strptime
In general using this function is a bad idea. for one, its *locale
dependent* which means any number of machines may perceive a given
string differently. this is okay-ish if you're reading your own
generated string, but certainly bad for parsing anything of a set
format.

this also removes the check and replacement for strptime, since
we obviously won't use it anymore.