audio/win32: add built-in directsound driver
I don't know if 32-bit audio works properly so I'm not even bothering
with setting it.

The waveout driver was renamed internally to say, well, `waveout`
instead of `win32` since it's no longer the only win32 backend

The pausing stuff uses a non-atomic variable, which shouldn't be a
problem since the audio thread only reads from it.