Clearer error messages when saving nameless samples
Currently the sample list's alt+o, alt+w, and alt-t operations fail with
"Sample <n> NOT saved! (File Error)" when the sample hasn't been given a
filename. With alt+w and alt+o the sample can have no filename if the
"filename" field has been left blank; with alt+t it happens when the
dialog text box is left blank.

Here I check for each of these operations that the filename is blank and
if so bail with a more descriptive error message.