sys/macos: lots of changes
the stupid config stuff inherited from SDL is completely gone now,
and replaced with values in our own config file. existing setups
probably broke because of this, if anyone actually knew about this.
whatever, the stuff that was in there was STUPID UGLY and I don't
wanna have to maintain it

in addition, now we ~should~ (crossing fingers intensely) have
proper Unicode paths, which means paths from before this are probably
also broken now if they include non-ASCII chars. Whatever.