vgamem: pack all values into 32-bits.
this is very hacky, but we kind of need to do it this way to
have speedy results on older systems (or maybe there's some other
bottleneck preventing it from running fast). for the most part
it looks a lot like the old code, but much more self-documented
instead of having random constants everywhere.