*: initial custom keybinds support code
this is very bare-bones as of now, but contains basically the bare
minimum to get things rolling. I've decided to hide all of the
implementation details within keybinds.c in an attempt to make
everything look "clean". Hopefully it doesn't have any sort of
taxing effect on performance! :')

I based the keybinds list in keybinds.h off of the original keybinds
PR since I didn't want to collect them all myself and wanted to focus
on making the external API the way I envisioned first