tests: refactor a bit, win32 is completely untested
I've changed the function parameters for os_exec to be a variable
argument function. This makes it look a lot more like the POSIX
execl. This also means we can take any number of arguments we want,
and (for the most part) all of the cruft can be kept inside of
the exec implementation. However, run_hook still needs to be
implemented separately.

os_shell is completely gone. There was no way to tell if the actual
program failed or the process creation failed. That ambiguity sucks,
especially when we want to fall back if the process creation failed.

Test output now uses a dynamic memory buffer instead of the linked
list shenanigans it was doing before. This is probably faster anyway.

I don't know if relying on argv[0] being the program name everywhere
is ok, but it's fine for now. It's okay on all the platforms that
matter. Plus, since we can fall back, who cares.