player: build proper AdLib VU meters
now it's actually based off of the output volume (yay!), and the
OPL2 source was updated to properly handle multi-channel output.
thus we no longer need a separate mixing buffer, nor special code
for the OPL2 branch, and everything can be consistent within
snd_fm.c.

OPLUpdateOne is now emulated by simply calling OPLUpdateMulti
with all of the buffers set to the same one. This shouldn't have
much of an impact on performance, but I suppose the real performance
hit is within emulating AdLib itself.