page_midiout: tell the format savers to not just throw out the midi data we've edited
this caused the .IT savers to just completely ignore the fact that
we've edited the midi data when creating a new file (whoops!!!)

the fact that this flag exists *at all* is questionable. for example,
what if we have a different default midi config? will it just not
get saved at all? this is incredibly fishy. it needs to get removed :)