page_midiout: turns out this is by design (??)
okay, so the midi output config is designed to be optional, and can
be enabled by the "Embed MIDI data" option, which annoyingly is NOT
on the MIDI output screen, but on the fucking MIDI ports screen.

annoying, and really bad UX design. why can't the toggle just be
on the output screen???? :(