midi-ip: make everything substantially more sane
now we're using the userdata as we should, and I got rid of most
of the stinky globals. at this point, all thats left is to move
all of the globals into midi_provider (where they belong).

it would also be nice to only allocate a file descriptor when we're
actually using it, but I have to go to work now.

other stuff might break, because I had to change some function parameters
to get userdata working right.