player: add building blocks for playback quirks
based off the OpenMPT code.

Saga Musix suggests to clear the quirks when the song is saved, so that
old incompatible modules don't "grandfather in" to newer versions. I think
that's definitely better than saving them (and possibly growing the resulting
file size), so that's what I'm going to implement

OpenMPT itself saves the state of the quirks. It is possible that we could
load and parse this data, so that they will play the same within both
players.