events: explicitly zero the entire event structure with memset
as C is a very good programming language with absolutely no footguns
anywhere, zero-initializing a union using `= {0}` isn't guaranteed
to actually zero out the entire thing (in fact clang DIDN'T do this
for a long while, and gcc was the only compiler it actually worked
right on. this seems to have been reversed recently though.)

so, replace all initializations of schism_event_t with memset.
there may be more union initializations that I'm missing, tho.