player: adopt OpenMPT defaults for volume ramping, fix weird clicking
the "weird clicking" in volume envelopes described in #591 was
actually entirely caused by a logic bug which seems to have been
fixed in OpenMPT. I've just applied it here, since it sounds much
nicer than what it was before.

also ramping values are shorter in general, but shouldn't introduce
much clicking; if it turns out to be a problem I can revert them,
but ideally these would be configurable by the user anyway.