fmt/aiff: save/load sample loops, volume, panning, etc
We also save the encoder name e.g. "Schism Tracker 20251008" in a
'APPL'+'pdos' chunk since it's the only one I found that seems to be
FOR saving the "application name".

We use another 'APPL' chunk with the identifier 'Schm' to store the
Modplug 'xtra' data, which has volume and panning info among other
things.

Also we store the song title now too which is cool