sdl[12/2]: set environment variables with SDL functions
this is damn annoying, but it makes things work right under
sdl2-compat.

SDL3 had the bright idea of grabbing the whole environment table
on startup. This means that to set an environment variable you
*absolutely must* call SDL_SetEnvironmentVariable. Unfortunately,
this has the side effect of breaking our SDL 1.2 and SDL 2 backends
because both of them relied on SDL reading from the current program
environment (which also relied on them using the same libc)

Now this basically makes our util function for setting an
environment variable temporarily entirely useless, since it's not
even used anywhere anymore. Punt.