config: honor gl_bilinear for old config files
this also properly sets gl_bilinear so older versions can read it.
this is really only relevant for pre-2022 builds, and if your
install is new enough, the config won't even be in the same
place on Linux. maybe if someone wants to run schism again after
a very long hiatus it will be useful.

it also sets gl_bilinear based on whether we're NOT using nearest
interpolation.

Also, "best" interpolation seems to make no difference, and even
looking through SDL 2 code there's literally nothing that takes
advantage of it (hence its removal in SDL 3), so we can probably
just make that an alias of linear scaling.