page_waterfall: make the FFT variable-width, many other things
for one, this makes the FFT tables allocated on the heap. to make
this less of a hit, we allocate them all in one batch. for debugging
purposes (i.e. memory overrun) there is a compile-time option to
make all variables their own allocation.

For some reason that I'm too lazy to look into, filling the
incoming buffers with `fft_bufsize` actually results in weird,
skewed results, and changing it to just `fft_size` results in
a much better result. Oops. (I really should figure out what's
REALLY going on)