fmt/s3m: make panning conversions not stink
as far as I know this is the best method of doing this. There's no
"best way" of converting IT panning to S3M panning, but this method
is the one that I think is the best option.

From what I've been told (I don't care enough to look into this
myself), IT uses this conversion algorithm, at least the one used
in s3m_pan_to_it().

hmm, maybe we should check for older Schism versions and use the
correct panning algorithm for each version. Or maybe it just
doesn't matter.